// graphql-client/src/features/auth/types/authuser.type.ts
import { ObjectType, Field } from 'type-graphql';
import { GraphQLJSONObject } from 'graphql-type-json';

@ObjectType()
export class AuthUserType {
  @Field()
  id: string;

  @Field()
  fId: string;

  @Field()
  firstName: string;

  @Field()
  lastName: string;

  @Field()
  username: string;

  @Field()
  email: string;

  @Field()
  mobile: string;

  @Field()
  password: string;

  @Field()
  isActive: boolean;

  @Field(() => [String], { nullable: true })
  roles: Array<any>;
}